+++ client/web/src/jscontext.ts
@@ -113,7 +113,8 @@ export interface SourcegraphContext extends Pick<Required<SiteConfiguration>, 'e
         brandName: string
     }
 
-    productResearchPageEnabled?: boolean
+    /** Whether the product research sign-up page is enabled on the site. */
+    productResearchPageEnabled: boolean
 
     /** The publishable key for the billing service (Stripe). */
     billingPublishableKey?: string