+++ schema/site_stringdata.go
@@ -895,8 +895,8 @@ const SiteSchemaJSON = `{
       "description": "Enables users access to the product research page in their settings.",
       "type": "boolean",
       "!go": { "pointer": true },
-      "default": true,
-      "group": "Misc."
+      "group": "Misc.",
+      "default": true
     }
   },
   "definitions": {